+++ src/App.jsx
@@ -9,6 +9,7 @@ import { privateRoutes, publicRoutes } from "../src//routes/routes";
 import "antd/dist/reset.css";
 import "react-toastify/dist/ReactToastify.css";
 import { useSelector } from "react-redux";
+import { LayoutWrapper } from "./components/LayoutWrapper";
 
 function App() {
   const user = useSelector((state) => state.user);
@@ -17,18 +18,22 @@ function App() {
     <>
       <div>
         <Router>
-          <Routes>
-            {publicRoutes.map((route, index) => (
-              <Route key={index} path={route.path} element={route.element} />
-            ))}
-            <Route
-              path="/"
-              element={<Navigate to={user ? "/dashboard" : "/login"} replace />}
-            />
-            {privateRoutes.map((route, index) => (
-              <Route key={index} path={route.path} element={route.element} />
-            ))}
-          </Routes>
+          <LayoutWrapper>
+            <Routes>
+              {publicRoutes.map((route, index) => (
+                <Route key={index} path={route.path} element={route.element} />
+              ))}
+              <Route
+                path="/"
+                element={
+                  <Navigate to={user ? "/dashboard" : "/login"} replace />
+                }
+              />
+              {privateRoutes.map((route, index) => (
+                <Route key={index} path={route.path} element={route.element} />
+              ))}
+            </Routes>
+          </LayoutWrapper>
         </Router>
       </div>
     </>
