+++ src/pages/Employees/EmployeeDetails.jsx
@@ -1,0 +1,12 @@
+import "../../styles/dashboard.css";
+import { EmployeeInfo } from "../../components/EmployeeInfo";
+
+const EmployeeDetails = () => {
+  return (
+    <div className="emp-info-div">
+      <EmployeeInfo />
+    </div>
+  );
+};
+
+export default EmployeeDetails;
