+++ src/styles/dashboard.css
@@ -18,14 +18,22 @@ background-color: #272626 !important;
   background-image: linear-gradient(to right, #b987f3, #5bd4ec);
   height: 40%;
 }
+:where(.css-dev-only-do-not-override-15rg2km).ant-menu-dark .ant-menu-item-selected, :where(.css-dev-only-do-not-override-15rg2km).ant-menu-dark>.ant-menu .ant-menu-item-selected {
+  color: #fff;
+  padding-top: 10px;
+}
 .ant-menu-item{
   color: white !important;
+  align-self: center;
   height: 90px;
 }
 .ant-menu-item:hover{
   color: white !important;
   background-color: #272626 !important;
   height: 40%;
+}
+.ant-layout{
+  font-size: large;
 }
 .logo {
   display: flex;
@@ -38,7 +46,20 @@ background-color: #272626 !important;
   max-height: 100%;
   align-self: center;
 }
+.ant-layout-sider-trigger{
+  display: none;
+}
 
+:where(.css-dev-only-do-not-override-15rg2km).ant-menu-inline.ant-menu-root .ant-menu-item > .ant-menu-title-content,
+:where(.css-dev-only-do-not-override-15rg2km).ant-menu-inline.ant-menu-root .ant-menu-submenu-title > .ant-menu-title-content {
+  display: flex; /* Set the display to flex */
+  align-items: center; /* Center the flex items vertically */
+}
+
+:where(.css-dev-only-do-not-override-15rg2km).ant-menu-dark .ant-menu-item-selected, :where(.css-dev-only-do-not-override-15rg2km).ant-menu-dark>.ant-menu .ant-menu-item-selected {
+  align-self: center;
+ }
+ 
  /* HEADER */
 
 .user-avatar {
@@ -74,7 +95,7 @@ background-color: #272626 !important;
 .emp-search {
   width: 65%;
   border-radius: 90px;
-  margin-left: 5%;
+  margin-left: 2%;
   z-index: 1;
 }
 
@@ -131,18 +152,19 @@ background-color: #272626 !important;
   color: white;
   width: 35%;
   height: 90%;
-  margin: 5px;
+  margin-right: 3%;
+  /* margin: 5px; */
   text-decoration: none;
   overflow: hidden;
   right: 0;
 }
 
 /* EMPLOYEE REGISTRATION */
-.content-div .emp-reg-container{
+.emp-reg-container{
   width: 50%;
   height: 95%;
   align-self: center;
-  background-color: #272626e6;
+  background-color: #272626;
   justify-self: center;
   border-radius: 30px;
   padding: 15px;
@@ -150,12 +172,14 @@ background-color: #272626 !important;
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  align-self: center;
 }
 
 .emp-reg-container .close-btn{
   color: gray;
   align-self: center;
 }
+
 .emp-reg-container .ant-input {
   background-color: black;
   border-color: gray;
@@ -181,6 +205,31 @@ background-color: #272626 !important;
   font-family: 'Courier New', Courier, monospace;
 }
 
+.emp-reg-btn-div{
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  width: 80%;
+}
+.emp-register-button {
+  background: linear-gradient(to left,#5bd4ec, #b987f3);
+  color: white;
+  width: 38%;
+  margin: 5px;
+  text-decoration: none;
+  overflow: hidden;
+  left: 0;
+}
+.emp-cancel-button {
+  background: linear-gradient(to left,gray, #272626);
+  color: white;
+  width: 38%;
+  margin: 5px;
+  text-decoration: none;
+  overflow: hidden;
+  left: 0;
+}
+
 /* EMPLOYEE LIST */
 
 .emp-list-div {
@@ -196,6 +245,7 @@ background-color: #272626 !important;
   background-color: #272626;
   padding: 5px;
   border: 1px solid #272626;
+  cursor: pointer;
 }
 
 .emp-card-div {
@@ -375,6 +425,7 @@ background-color: #272626 !important;
   color: white;
   height: 30px;
 }
+
 
 
 @media screen and (max-width: 600px) {
